modules/tool-stitch-panels/pyproject.toml
[project]
name = "amplifier-module-tool-stitch-panels"
version = "0.1.0"
description = "Stitch multiple PNG panel images into a single vertically stacked infographic"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = ["pillow"]

[project.entry-points."amplifier.modules"]
tool-stitch-panels = "amplifier_module_tool_stitch_panels:mount"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["amplifier_module_tool_stitch_panels"]